Added SimulatedProtein - Lifted from one of Austin's branches, will need to be refined down to what is essential to keep
Added MS1DatabaseParser - create a protein list from database as simulated proteins

Added MS1SearchEngine -  Gets constructed with a database and a tolerance (20 ppm by default). PeakScorer() method will take the scan, deconvolute it, and check for the monoisotopic mass within the list of database proteins. If the mass is found within the database, the corresponding ScoreTable entry is incremented. This means a score of zero is unrecognized

RealTimeSampleProcessingExample - updated to be constructed with a MS1DatabaseParser and MS1Search engine. It will receive a scan from the instrument via an event, run it through the search engine, and then grab the top 5 peaks (by intensity) not found within the database

TestMethods - First is simple and just allows me to enter the code quickly in debug mode, ignore for the time being.  Second one takes a search result from metamorpheus where there were no PTM's, pulls out a third of the proteins found, and feeds the rest into the program as a database. It then compares the peaks the program selected to those that were removed. Returns true if at least 70% of the removed proteins have their monoisotopic masses selected for fragmentation throughout the program. The test was returning true with 90% hit rate before the file had to be shrunk down to size.

Other minor changes
Added a summary to a ppmTolerance method to help me
Added in AminoAcidPolymer constructor to work from a protein object
In SSRCalc3. generalized the ScoreSequence method to use a string so it can be called from any object with a sequence. Created an overload to not break any previous implementations.